requirements.txt
requests
axios
python-dotenv
pyyaml